+++ src/commands/TextEdits.ts
@@ -23,11 +23,6 @@ export class TextEdits {
       return;
     }
 
-    if (selectionText.length < 0 || selectionText.length > 1000) {
-      window.showErrorMessage("Please select a text between 1 and 1000 characters.");
-      return;
-    }
-
     window.withProgress({
       location: ProgressLocation.Notification,
       cancellable: false,