+++ src/constants/ExtensionConstants.ts
@@ -1,0 +1,5 @@
+
+
+export const ExtensionConstants = {
+  maxChars: 2500,
+};