+++ src/providers/index.ts
@@ -1,0 +1,2 @@
+export * from './CommandTreeviewProvider';
+export * from './CommandsWebviewProvider';