+++ src/service/index.ts
@@ -1,1 +1,3 @@
 export * from './AiService';
+export * from './Extension';
+export * from './Logger';