+++ src/utils/getSelectedText.ts
@@ -1,4 +1,5 @@
 import { window } from 'vscode';
+import { ExtensionConstants } from '../constants';
 
 export const getSelectedText = () => {
   let editor = window.activeTextEditor;
@@ -8,5 +9,11 @@ export const getSelectedText = () => {
   }
 
   const selectionText = editor.document.getText(editor.selection);
+
+  if (selectionText.length < 0 || selectionText.length > ExtensionConstants.maxChars) {
+    window.showErrorMessage(`Please select a text between 1 and ${ExtensionConstants.maxChars} characters.`);
+    return;
+  }
+
   return selectionText;
 }